django_vue/requirements.txt
django-cors-headers==3.13.0
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.2.0
stripe==3.5.0
django-imagekit==4.1.0